Added list.id to thought.list.id error checking

+++ idealist.rb
@@ -20,14 +20,14 @@ DataMapper.setup(:default, {
 class Thought
   include DataMapper::Resource
 
-  property :id, Serial    # primary serial key
+  property :id, Serial
   property :body, Text
   property :created_at, DateTime
   property :updated_at, DateTime
 
   validates_length_of :body, :minimum => 1
 
-  belongs_to :list, :required => false
+  belongs_to :list, :required => true
 
   def to_json(*a)
     JSON.pretty_generate({:id => id, :body => body, :created_at => created_at, :updated_at => updated_at})
@@ -71,9 +71,12 @@ post '/api/lists/?' do
     list = List.new(:name => data['name'])
     list.created_at = Time.now
     list.updated_at = Time.now
-    list.save
-    status 201
-    list.to_json
+    if list.save
+      status 201
+      list.to_json
+    else
+      status 412
+    end
   end
 end
 
@@ -87,9 +90,12 @@ put '/api/lists/:id/?' do
     else
       list.name = data['name']
       list.updated_at = Time.now
-      list.save
-      status 200
-      list.to_json
+      if list.save
+        status 200
+        list.to_json
+      else
+        status 412
+      end
     end
   else
     status 404
@@ -128,9 +134,12 @@ post '/api/lists/:listId/thoughts/?' do
       thought = Thought.new(:list_id => list.id, :body => data['body'])
       thought.created_at = Time.now
       thought.updated_at = Time.now
-      thought.save
-      status 201
-      thought.to_json
+      if thought.save
+        status 201
+        thought.to_json
+      else
+        status 412
+      end
     end
   else
     status 404
@@ -141,16 +150,19 @@ end
 put '/api/lists/:listId/thoughts/:id/?' do
   list = List.get(params[:listId])
   thought = Thought.get(params[:id])
-  if list && thought
+  if list && thought && thought.list.id === list.id
     data = JSON.parse(request.body.read.to_s)
     if data.nil? || !data.has_key?('body')
       status 400
     else
       thought.body = data['body']
       thought.updated_at = Time.now
-      thought.save
-      status 201
-      thought.to_json
+      if thought.save
+        status 200
+        thought.to_json
+      else
+        status 412
+      end
     end
   else
     status 404
@@ -161,7 +173,7 @@ end
 delete '/api/lists/:listId/thoughts/:id/?' do
   list = List.get(params[:listId])
   thought = Thought.get(params[:id])
-  if list && thought
+  if list && thought && thought.list.id === list.id
     thought.destroy
     status 200
   else
@@ -182,30 +194,32 @@ end
 
 post '/lists/create/?' do
   @list = List.new(:name => params[:list_name])
-  @list.save
-  redirect '/lists'
+  if @list.save
+    redirect '/lists/'
+  else
+    redirect '/lists/new/'
+  end
 end
 
 get '/lists/edit/:id/?', :provides => 'html'  do
   @list = List.get(params[:id])
-  if @list
-    haml :"lists/edit"
-  else
-    redirect '/lists'
-  end
+  haml :"lists/edit"
 end
 
 post '/lists/update/:id/?' do
   @list = List.get(params[:id])
   @list.name = params[:list_name]
-  @list.save
-  redirect '/lists'
+  if @list.save
+    redirect '/lists/'
+  else
+    redirect "/lists/edit/#{@list.id}/"
+  end
 end
 
 delete '/lists/:id/?' do
   @list = List.get(params[:id])
   @list.destroy
-  redirect '/lists'
+  redirect '/lists/'
 end
 
 get '/lists/:listId/thoughts/?', :provides => 'html'  do
@@ -222,43 +236,53 @@ end
 post '/lists/:listId/thoughts/create' do
   @list = List.get(params[:listId])
   @thought = Thought.new(:list_id => params[:listId], :body => params[:thought_body])
-  redirect "/lists/#{@list.id}/thoughts"
+  if @thought.save
+    redirect "/lists/#{@list.id}/thoughts/"
+  else
+    redirect "/lists/#{@list.id}/thoughts/new/"
+  end
 end
 
 get '/lists/:listId/thoughts/:id/?', :provides => 'html' do
   @list = List.get(params[:listId])
   @thought = Thought.get(params[:id])
-  if @thought
+  if @thought.list.id === @list.id
     haml :"thoughts/show"
   else
-    redirect "/lists/#{@list.id}/thoughts"
+    redirect "/lists/#{@list.id}/thoughts/"
   end
 end
 
 get '/lists/:listId/thoughts/edit/:id/?', :provides => 'html' do
   @list = List.get(params[:listId])
   @thought = Thought.get(params[:id])
-  if @thought
+  if @thought.list.id === @list.id
     haml :"thoughts/edit"
   else
-    redirect "/lists/#{@list.id}/thoughts"
+    redirect "/lists/#{@list.id}/thoughts/"
   end
 end
 
 post '/lists/:listId/thoughts/update/:id/?' do
   @list = List.get(params[:listId])
   @thought = Thought.get(params[:id])
-  @thought.body = params[:thought_body]
-  if @thought.save
-    redirect "/lists/#{@list.id}/thoughts/#{@thought.id}"
+  if @thought.list.id === @list.id
+    @thought.body = params[:thought_body]
+    if @thought.save
+      redirect "/lists/#{@list.id}/thoughts/#{@thought.id}/"
+    else
+      redirect "/lists/#{@list.id}/thoughts/edit/#{@thought.id}/"
+    end
   else
-    redirect "/lists/#{@list.id}/thoughts"
+    redirect "/lists/#{@list.id}/thoughts/"
   end
 end
 
 delete '/lists/:listId/thoughts/:id/?' do
   @list = List.get(params[:listId])
   @thought = Thought.get(params[:id])
-  @thought.destroy
-  redirect "/lists/#{@list.id}/thoughts"
+  if @thought.list.id === @list.id
+    @thought.destroy
+  end
+  redirect "/lists/#{@list.id}/thoughts/"
 end
